Use `arburg` instead of `ar` for autoregression
`ar` is only available with the System Identification Toolbox, while `arburg` is available in the Signal Processing Toolbox (which is available in all Student Editions of MATLAB). I have swapped `arburg` for `ar`, keeping the same coefficient order. This does, however, use Burg's lattice-based method for computing the model, instead of the forward-backward approach that was used in `ar`.